Switch back to manual coverage.py parser
The new parser fails for many people on CI because it requires a
command-line tool that is proving tricky to guarantee installed and, if
installed, tricky to guarantee we can find it.

As a stopgap, revert that change and push that release out. The work in
the commit is good, but we need to focus more on ensuring access to the
command-line tool.

This reverts commit ac469acf9bbc239ef70900d87bdef7689a5dd488. Closes #119