Reinstate updated python parser (#122)
This reinstates the work done in ac469acf that was rolled back in 15f4e78.
We found that users sometimes had a .coverage file, but did not have the
coverage program installed, causing an error to be raised.

These changes add a check for the coverage executable in the matches?
method for the python parser. If a file matches the python format, then we
also do a secondary check to make sure that coverage is installed. If not,
then we output a message about the missing program when the debug flag
is passed.

If a user specifies format: python but does not have coverage installed
this will raise an error with more specific information about how to solve
the issue.
    
---------
    
Co-authored-by: Mike Burns <mburns@thoughtbot.com>