Format forcing does not bypass valid file check (#125)
When forcing a format, the system still collects all possible
coverage files across all file types, and passes them directly
to the specified parser, bypassing the parser's matches? method
that checks to make sure that the file is of a valid format. This
is causing issues when a user is forcing the cobertura format, 
but has a .coverage file without the coverage executable installed.

These changes ensure that if a user specifies a format, then
only the files that match the glob-pattern for the specified
parser will be passed to the parser. In addition, the matches?
method is no longer bypassed, which ensures that even if a
file matches the correct naming pattern, there will be no attempt
to parse it if the actual file format is incorrect.

If a user specifies the python format for a .coverage file, but does
not have the coverage executable installed, then an error will be
raised. If no format is specified, and the user has a .coverage file
but does not have the coverage executable installed, then a debug
warning message only will be displayed.