Remove restriction on file extension for some parsers
Problem
--------
All files are now run through the matches?
method, which is causing issues for the lcov,
gcov, and simplecov parsers since the
matches? method checks that files have a
particular file extension. This is causing failures
when users pass a custom file.

Solution
--------
Always return true for the matches? method
for lcov, gcov, and simplecov parsers. If the
user does not specify a custom file, then
only files that match the extensions listed in
this method will be returned anyway due to the
globs method. If a user specifies a custom
file, we do not want to exclude the file just
because it does not have the correct extension.